+++ samples/compose-viewmodels/screen-details/src/commonMain/kotlin/dev/zacsweers/metro/sample/composeviewmodels/details/DetailsScreen.kt
@@ -5,6 +5,7 @@ package dev.zacsweers.metro.sample.composeviewmodels.details
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
@@ -25,7 +26,7 @@ fun DetailsScreen(
     assistedMetroViewModel<DetailsViewModel, DetailsViewModel.Factory> { create(data) },
 ) =
   Column(
-    modifier = modifier,
+    modifier = modifier.fillMaxSize(),
     horizontalAlignment = Alignment.CenterHorizontally,
     verticalArrangement = Arrangement.Center,
   ) {