+++ samples/compose-viewmodels/screen-details/src/commonMain/kotlin/dev/zacsweers/metro/sample/composeviewmodels/details/DetailsViewModel.kt
@@ -8,7 +8,6 @@ import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedFactory
 import dev.zacsweers.metro.AssistedInject
 import dev.zacsweers.metro.ContributesIntoMap
-import dev.zacsweers.metro.binding
 import dev.zacsweers.metrox.viewmodel.ManualViewModelAssistedFactory
 import dev.zacsweers.metrox.viewmodel.ManualViewModelAssistedFactoryKey
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -31,7 +30,7 @@ class DetailsViewModel(@Assisted val data: String) : ViewModel() {
 
   @AssistedFactory
   @ManualViewModelAssistedFactoryKey(Factory::class)
-  @ContributesIntoMap(AppScope::class, binding<ManualViewModelAssistedFactory>())
+  @ContributesIntoMap(AppScope::class)
   fun interface Factory : ManualViewModelAssistedFactory {
     fun create(@Assisted data: String): DetailsViewModel
   }